@@ -2,7 +2,7 @@
   "name": "chc-integral",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^7.2.0",
-    "@angular/core": "^7.2.0"
+    "@angular/common": "^8.2.14",
+    "@angular/core": "^8.2.14"
   }
 }
