@@ -1,20 +1,10 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
-  selector: 'spv-root',
+  selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+  styleUrls: ['./app.component.less']
 })
-export class SpvAppComponent implements OnInit {
+export class AppComponent {
   title = 'spnview';
-  @Input()
-  pool: string;
-  area: string;
-
-   ngOnInit() {
-       const el = window.document.querySelector('spv-root');
-       this.pool = el && el.attributes["pool"] && el.attributes["pool"].value;
-       this.area = el && el.attributes["area"] && el.attributes["area"].value;
-       window.console.log(el);
-   }
 }
